requirements.txt
Flask==2.0.3
mysql-connector==2.2.9
configparser==5.0.1
